requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
yt-dlp>=2024.1.0
scenedetect[opencv]>=0.6.4
aiofiles>=23.2.1
python-multipart>=0.0.6
